4.2.4 refactoring movie-tabs.jsx, add new component movie-tabs.jsx
